requirements.txt
plotly
streamlit
pandas
numpy
psycopg2-binary
matplotlib
hdbscan
scikit-learn
seaborn